wip! API endpoints for managing identifier sets and minting identifiers
Currently unused, but with a web interface for minting new identifiers
in mind.

I wrote these endpoints while integrating identifier minting and barcode
label generation, only to end up refactoring the meat of it out and
generating labels via a CLI command instead.  We do eventually want a
web frontend for this, so I'm going to commit these as-is despite them
not being used so that the work doesn't disappear.

Also also, read-only barcode access for Audere?